Updated functions and quiz questions
Updated the functions for each part to make better use of outputs, nested functions and to make sure data that was being used is numeric and not as character.
Added the on.exit() function to change the WD.
For whatever reason when I went to assign the corr function as an object, it saved the object as the setwd() string = directory argument and no other data even though it would spit out proper outputs. return() was also utilized and allowed for complete() function to be used within the corr() function easily - makes for much easier work.