Updated directory argument
Updated directory argument to better match the coursera code. I originally had getwd(directory) set in the function and made specdir<-"C:/Users/...." as the full location.
To get the wd properly set, specdir had to be ran prior to the function being run and specdir used as the directory argument as
opposed to "specdata"